schema now made similar to tpch spec

+++ scala/src/ppl/dsl/optiql/baseline/benchmarks/tpch/schema/Customer.scala
@@ -10,7 +10,7 @@ class Customer (
   val c_address: String,
   val c_nationkey: Int,
   val c_phone: String,
-  val c_acctbal: Float,
+  val c_acctbal: Double,
   val c_mktsegment: String,
   val c_comment: String
 )
