+++ scala/src/ppl/dsl/optiql/baseline/benchmarks/tpch/schema/Order.scala
@@ -8,7 +8,7 @@ class Order (
   val o_orderkey: Int,
   val o_custkey: Int,
   val o_orderstatus: Char,
-  val o_totalprice: Float,
+  val o_totalprice: Double,
   val o_orderdate: Date,
   val o_orderpriority: String,
   val o_clerk: String,
