+++ scala/src/ppl/dsl/optiql/baseline/benchmarks/tpch/schema/PartSupplier.scala
@@ -7,7 +7,7 @@ class PartSupplier (
   val ps_partkey: Int,
   val ps_suppkey: Int,
   val ps_availqty: Int,
-  val ps_supplycost: Float,
+  val ps_supplycost: Double,
   val ps_comment: String
 )
